CommonService/requirements.txt
sqlalchemy
quart
asyncmy
hypercorn
python-dotenv
mysqlclient
pymysql
cryptography